refactor(atelier_dom): split lib into cohesive modules
